Voltage Indicator
To get voltage of two 12 volt batteries. Using C code within a
PIC16f690, the user pushes a button in order to display each voltage on
a 16x2 LCD display. The code is configured to interface with the LCD
using 8 bits, 4 "high" bits, or 4 "low" bits, depending on what pins
you'd like open for use.